acme-client: stop reaching into X509
Prepare for an upcoming change in libcrypto and retrieve the stack
of extensions via X509_get0_extensions(). Simplify the for loop by
relying on the fact that empty or NULL stacks have an sk_num() of 0
and -1, respectively, so the loop won't be entered and the extsz
dance is unnecessary.

ok florian

(cherry picked from commit 5fc393520e66be4c7d4cf4501e6eba680d3d5606)