Make function prototype parsing a bit stricter, reducing the risk of accidental misparsing:
Require whitespace after the function return type (before the
asterisk indicating that the function returns a pointer, if any)
and do not accept whitespace between the function name and the
opening parenthesis of the parameter list.  These changes are not
a problem because we want that style for KNF reasons anyway.

(cherry picked from commit 3958557a61916ede4b245906db047648e7f25d05)