Simplify DH_check_params a bit.
It makes no sense to allocate an entire BN_CTX if we only use it to
get a single BIGNUM, from which we subtract 1 to compare it to g.
We can just use a plain BIGNUM and delete a bunch of lines.

ok inoguchi jsing

(cherry picked from commit adfe9e36779c560e6c43d13cf305fabc6719e9d4)