Add aplsmc(4), a driver for the SMC found on Apple M1 SoCs.
The SMC implements a lot of functionality.  For now the driver only
implements a bunch of sensors.  This is a small subset of the sensors
that are made available by the SMC as we don't know what measurments
are provided for most sensors.

ok patrick@

(cherry picked from commit d80548ace53d64d712ae062359a16cdad3988e24)