Simplify tlsext_supported_groups_server_parse
Add an early return in the s->internal->hit case so that we can unindent
a lot of this code. In the HRR case, we do not need to check that the list
of supported groups is unmodified from the first CH. The CH extension
hashing already does that for us.

ok jsing

(cherry picked from commit 90dae725fcf59f84132aec26dc3dd569b667c283)