Ignore error when we try to delete an address that's already gone.
This will happen when an address expires because the vltime drops to
zero. The kernel then deletes the address and slaacd tries to do so,
too. The correct fix is to track in slaacd that the kernel already
deleted the address for us, but that's too much work shortly before a
release so just hide the ugly warning for now, it's harmless.
Problem reported by semarie some time ago.
OK deraadt, benno

(cherry picked from commit 51e6ed8b7e7efec0d50fadf56450317d04e1bb47)