Length check URI before strncasecmp()
A priori URI is not NUL terminated, so we should first check it is long
enough before comparing it against proto. As a side effect, this now
rejects "https://" and "rsync://", which are invalid due to the missing
host in the authority section.

ok claudio

(cherry picked from commit 3b405428c938f2c0c4fa4849c56380774b3c1043)