further speed up delivery of interrupts to a running vcpu.
this records which physical cpu a vcpu is running on. this is used
by the code that marks a vcpu as having a pending interrupt to check
if the vcpu is currently running. if it thinks the vcpu is running,
it sends a nop IPI to the physical cpu it is running on to trigger
a vmexit, which in turn runs interrupt handling in the guest.

ok mlarkin@

(cherry picked from commit 0d8b5945bad2cf707e51761352bff6611f56316b)