Add qcrtc(4), a driver for the RTC found on Qualcomm PMICs.  Unfortunately on the Lenovo x13s it does not seem to contain the offset from 1970.  Also we are not allowed to write to the RTC.
I was told that the offset I need to calculate the actual date is sorted in
an UEFI variable, which can be accessed over a secure channel.  So as long
as we don't have that channel, this driver is useless.  Good start though.

ok kettenis@

(cherry picked from commit ebbbade523ef9055672aa4e81615aaec8377f8f3)