acme-client: encode unexpected SANs before printing
If a SAN isn't configured, it could be anything, so make printing it safe
using strvisx(). If it is configured but duplicate, printing it should be
fine, so don't bother. This removes two XXX added in the previous commit.

ok florian

(cherry picked from commit 0b60193df276dc6f43ebb60edd076225d1cc0850)