Fix a -0 corner case in BN_div_internal()
If the numerator is negative, the numerator and divisor are the same
length (in words) and the absolute value of the divisor > the absolute
value of the numerator, the "no_branch" case produces -0 since negative
has already been set. Call BN_set_negative() at the end of the function
to avoid this.

ok tb@

(cherry picked from commit ae57c522841642bd680c4788b90ac931f9512451)