Improve media disk question default
Sets on an unmounted disk partition are most likely to be on install media,
e.g. CD or USB stick.

Default 'Which disk contains the install media?' to the first disk that is
a) not the root disk and b) not a disk with softraid chunks (hosting the
root disk, for example).

This makes it point at the right disk in most setups.
All disks remain valid answers, this is only about the default.

OK afresh1

(cherry picked from commit a764052d3bba9866cb0b79d83a539dc876d2c7be)