drop needless global tagsfn path
mg kept the path to the last loaded tag file in tagsfn which was used
for both the lazy loading (now removed) and as a flag to know if any
tags are currently loaded.  It's redundant and complicates the code:
just check if the tree is empty instead.

ok tb@

(cherry picked from commit 58c3a53a57fe3727a59a3b5be167c18302b714a2)