Rework BN_exp() a bit
This mostly only cleans up the mess that it was - which doesn't stand out
because of the horror that lurks in the rest of this file. It avoids
copying the partial calculation out on error and does away with some
other weirdness.

with/ok jsing

(cherry picked from commit 8eb14e88c92da3bec2a523de39fca075a652b8f0)