support Designware Mobile Storage Host Controllers
this is used on rk356x and rk3588 SoCs as the emmc controller.

this is actually an sdhc(4) controller, but it has a lot of extra
registers and code for dealing with clock rates. i could have added
this to sdhc_fdt.c, but the amount of this extra code and the lack
of overlap with other chips meant id be splitting the code up based
on the chip id, which is something the autoconf/device framework
already does for us.

some help and ok from patrick@
ok jmatthew@

(cherry picked from commit 7d57b1b5686200292ec1796659905a3558bac679)