Provide X509_VERIFY_PARAM_set_hostflags()
This is needed for an upcoming regress test that needs to access the
hostflag. This is public API in OpenSSL but since nothing seems to be
using this, this accessor will be kept internal-only for the time being.

ok jsing

(cherry picked from commit e51d4db3dada5e138401dc8d637c3ce8a97cb983)