Add a BN_R_INVALID_ARGUMENT error code
One problem with OpenSSL error codes is that they tend to be too specific
(another problem is that they are extremely ugly). So add an EINVAL-style
error code. This will be used in an upcoming commit to disallow aliasing
of the 'return value' with the modulus in BN_mod_* functions and should
be applicable elsewhere, outside of this one narrow use case.

ok jsing

(cherry picked from commit 3c52cd979fdcef9aada458e7d2c616510355121b)