Avoid undefined behavior in tlsexttest
An empty array cannot be initialized and it is an incomplete type, so the
sizeof() operator can't be applied to it. Therefore initialize it with a
zero byte and replace sizeof() use with lengths.

Slightly tweaked version of a patch by inoguchi in portable.
Discussed with bcook, beck, kettenis

(cherry picked from commit 1964a63a33254fd27b37b5f191d078dccb4f0c6d)