Add support for the RK3588 PCIe3 PHY to rkpciephy(4).  The PHY controls 4 lanes that can be routed to 4 of the 5 PCIe controllers in the following configurations.
Either all 4 lanes go to the x4 controller, then the other 3 controllers
stay disabled.  If those are split, lane 0 goes to the x4 controller and
lane 2 goes to the x2 controller.  Lane 1 now either goes to the x4 one
to build a x2, or it gets routed to a x1 controller.  Lane 3 also either
goes to the x2 one to build a x2, or it gets routed to a x1 controller.

If lanes 1 or lane 3 are configured, they are routed towards the PCIe3 PHY
instead of the Combo PHY.

ok kettenis@

(cherry picked from commit be60217fefd91228fe62b47903878687da42b03d)