Import EVP_MD_meth_new.3 from OpenSSL 1.1.1 head
As usual, this has a lot of room for improvement, but it is better than
nothing at all. It leaves a dangling reference to EVP_MD_CTX_md_data(),
which I leave for schwarze to document. This is a terrible API and the
way it is used in the wild is terrifying.

(cherry picked from commit ef7b131135f1069c37993b8a7bc5f231f085a18c)