Prepare the ssltest for the upcoming bump
This is a hack. The test is in rather poor shape and it is hard to tell
whether it still does what it is supposed to be doing. Hopefully somemone
will rewrite this in a style that doesn't make me squeal on opening this
file...

(cherry picked from commit 2287d13a721783c23c3036e2df45e2d342fd91bb)