Remove ASN1_BIT_STRING_check
This was added with the TS code for no discernible reason. I could not
find a single consumer. In the unlikely event that you need this, it is
easy enough to write a better version of it yourself.

ok jsing

(cherry picked from commit 1411e9ef2a4ec93e44d8b2cd207dd80b45d88efe)